requirements.txt
bleach==2.1.3
boto3==1.6.8
botocore==1.9.8
dj-database-url==0.5.0
Django==2.0.3
django-heroku==0.3.1
django-rest-framework==0.1.0
django-storages==1.6.5
djangorestframework==3.7.7
docutils==0.14
gunicorn==19.7.1
html5lib==1.0.1
jmespath==0.9.3
Pillow==5.0.0
psycopg2-binary==2.7.4
python-dateutil==2.6.1
pytz==2018.3
s3transfer==0.1.13
six==1.11.0
webencodings==0.5.1
whitenoise==3.3.1
